Create problem 2
PROBLEM:
An automobile company has serial number for engine parts starting from AA0 to FF9. The other characteristics of parts are year of manufacture, material and quantity manufactured.
(1) Specify a structure to store information corresponding to a part.
(2) Write a program to retrieve information on parts with serial numbers between BB1 and CC6.